EliteTruthEngine/requirements.txt
fastapi
uvicorn
llama-cpp-python
txtai
sentence-transformers
